chore: move buffer tests
